refactor: give Polecat.AspNetCore.Testing a factory instead of an entry point
Preparation for the xUnit v3 migration, split out so that migration stays a
mechanical package swap.

The app under test lived in top-level statements in Program.cs, and the six
Alba suites bootstrapped it with AlbaHost.For<Program>() — WebApplicationFactory,
which works by invoking the assembly's entry point. That arrangement cannot
survive xUnit v3: a v3 test assembly is a self-executing test runner, so it has
to own Main. With top-level statements present the generated runner entry point
loses, and the test executable boots Kestrel instead of discovering tests.

This is not speculative. Running the v3 conversion against this project first
reported zero tests on both target frameworks:

  [xUnit.net] Exception discovering tests: System.InvalidOperationException:
  Test process did not respond within 60 seconds
     at Xunit.v3.TestProcessLauncherAdapter.GetAssemblyInfo(...)
  No test is available in Polecat.AspNetCore.Testing.dll.

Program.cs becomes TestApp.cs: a static CreateBuilder()/Configure(app) pair
holding the identical service registration and the identical fourteen endpoint
mappings. Nothing in the assembly claims an entry point anymore. The suites
switch to AlbaHost.For(TestApp.CreateBuilder(), TestApp.Configure), an overload
Alba has offered since 7.x precisely so a minimal-API host can be built without
WebApplicationFactory. The file is renamed rather than kept as Program.cs
because a Program.cs with no Program in it is a trap for the next reader.

NoWarn CS7022 goes away with it — that suppression existed only because
Microsoft.NET.Test.Sdk's generated Main and the top-level statements were
fighting over the entry point.

Still on xUnit v2 here, so this is provably behavior-preserving on its own:
65 tests, all passing, same count as before the change.
