+++ src/Polecat.AspNetCore.Testing/Polecat.AspNetCore.Testing.csproj
@@ -3,7 +3,6 @@
     <PropertyGroup>
         <IsPackable>false</IsPackable>
         <NoDefaultLaunchSettingsFile>true</NoDefaultLaunchSettingsFile>
-        <NoWarn>CS7022</NoWarn>
     </PropertyGroup>
 
     <ItemGroup>
